Fixed case testing
Most of the run.py files for the tutorials encountered a TabError following the Python 3 compatibility change. This fixes this error (mostly changes TAB+space to just TAB). There were also some minor errors with a few of the case tests due to using floats where there should be integers, which are fixed here as well.